UI (#1240)
* Remove @grouparoo/ui

* Create @grouparoo/ui-enterprise

* Updating things automatically

* add missing log pages

* Shared Components

* WIP: Pages

* Shared pages

* 2 staging servers and EnterpriseLink

* udpate `grouparoo init` package template

* testing for 3 ui packages

* Update webpack config to allow for compilation in ui-components

* re-build lockfile

* test staging servers and UI packages on CI

* update licence checker and cache paths

* public orginization

* fix ui test paths on CI

* cache next builds

* remove un-used deps

* fixup lockfile

* Try axios instaead of isomorphic-fetch

* set global.URL in Jest setup

* udpate lockfile

* organize depenencies

* ui libs need core

* rebase with master + next 10.0.6

* license updates

* styled-jsx needed on server

* open docs CTA in new tab

* hide setup step CTAs for community eddition

* conditional links for Run and Export details pages

* test setup card CTA only for enterprise eddtion

* start a run from source list

* Variable app in procfile

+++ core/__tests__/actions/files.ts
@@ -33,17 +33,13 @@ describe("actions/files", () => {
 
     beforeAll(async () => {
       const file = "/tmp/roo.png";
-      if (await fs.pathExists(file)) {
-        fs.unlinkSync(file);
-      }
+      if (await fs.pathExists(file)) fs.unlinkSync(file);
 
       await fs.copy(
         path.join(
           __dirname,
           "..",
           "..",
-          "..",
-          "ui",
           "public",
           "images",
           "logo",
